+++ packages/contracts/contracts/test/invariant/helpers/OperationGenerator.sol
@@ -23,6 +23,7 @@ import {ActorSumSet, LibActorSumSet} from "../helpers/ActorSumSet.sol";
 import {LibDepositRequestArray} from "../helpers/DepositRequestArray.sol";
 import {Utils} from "../../../libs/Utils.sol";
 import {AssetUtils} from "../../../libs/AssetUtils.sol";
+import {InvariantUtils} from "../helpers/InvariantUtils.sol";
 import "../../../libs/Types.sol";
 
 struct GenerateOperationArgs {
@@ -49,7 +50,7 @@ struct GeneratedOperationMetadata {
     bool[] isSwap;
 }
 
-contract OperationGenerator is CommonBase, StdCheats, StdUtils {
+contract OperationGenerator is InvariantUtils {
     uint256 constant ERC20_ID = 0;
     uint256 constant DEFAULT_EXECUTION_GAS_LIMIT = 2_000_000;
     uint256 constant DEFAULT_PER_JOINSPLIT_VERIFY_GAS = 220_000;
@@ -80,13 +81,13 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
 
         // Get random args.joinSplitPublicSpends
         uint256[] memory joinSplitPublicSpends = _randomizeJoinSplitAmounts(
-            args.seed,
+            _rerandomize(args.seed),
             totalJoinSplitUnwrapAmount
         );
 
         // Get random numActions using the bound function, at least 2 to make space for token
         // approvals in case of a swap
-        uint256 numActions = bound(args.seed, 2, 5);
+        uint256 numActions = bound(_rerandomize(args.seed), 2, 5);
 
         uint256 gasToReserve = _opMaxGasAssetCost(
             DEFAULT_PER_JOINSPLIT_VERIFY_GAS,
@@ -115,12 +116,19 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
             ERC20_ID
         );
 
+        uint256 gasAssetRefundThreshold = bound(
+            _rerandomize(args.seed),
+            0,
+            totalJoinSplitUnwrapAmount
+        );
+
         FormatOperationArgs memory opArgs = FormatOperationArgs({
             joinSplitToken: args.joinSplitToken,
             gasToken: args.gasToken,
             root: args.root,
             joinSplitPublicSpends: joinSplitPublicSpends,
             encodedRefundAssets: encodedRefundAssets,
+            gasAssetRefundThreshold: gasAssetRefundThreshold,
             executionGasLimit: DEFAULT_EXECUTION_GAS_LIMIT,
             maxNumRefunds: DEFAULT_MAX_NUM_REFUNDS, // TODO: take based on number of swaps
             gasPrice: compensateBundler ? 1 : 0,
@@ -145,8 +153,13 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
             } else {
                 // Overflow here doesn't matter given all we need are random nfs
                 unchecked {
-                    _op.joinSplits[i].nullifierA = args.seed + (2 * i);
-                    _op.joinSplits[i].nullifierB = args.seed + (2 * i) + 1;
+                    _op.joinSplits[i].nullifierA =
+                        _rerandomize(args.seed) +
+                        (2 * i);
+                    _op.joinSplits[i].nullifierB =
+                        _rerandomize(args.seed) +
+                        (2 * i) +
+                        1;
                 }
             }
         }
@@ -182,7 +195,7 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
                     args.exceedJoinSplitsMarginInTokens;
             }
             uint256 transferOrSwapAmount = bound(
-                args.seed,
+                _rerandomize(args.seed),
                 0,
                 transferOrSwapBound
             );
@@ -264,11 +277,15 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
         );
 
         uint256 swapErc20OutAmount = bound(
-            args.seed,
+            _rerandomize(args.seed),
             0,
             type(uint256).max - args.swapErc20.totalSupply()
         );
-        uint256 swapErc1155OutAmount = bound(args.seed, 0, 10_000_000);
+        uint256 swapErc1155OutAmount = bound(
+            _rerandomize(args.seed),
+            0,
+            10_000_000
+        );
         SwapRequest memory swapRequest = SwapRequest({
             assetInOwner: address(args.handler),
             encodedAssetIn: encodedAssetIn,
@@ -290,14 +307,18 @@ contract OperationGenerator is CommonBase, StdCheats, StdUtils {
     function _randomizeJoinSplitAmounts(
         uint256 seed,
         uint256 totalAmount
-    ) internal view returns (uint256[] memory) {
+    ) internal returns (uint256[] memory) {
         uint256 numJoinSplits = bound(seed, 1, 5); // at most 5 joinsplits
         uint256[] memory joinSplitAmounts = new uint256[](numJoinSplits);
 
         uint256 remainingAmount = totalAmount;
         for (uint256 i = 0; i < numJoinSplits - 1; i++) {
             // Generate a random amount for the current join split and update the remaining amount
-            uint256 randomAmount = bound(seed, 0, remainingAmount);
+            uint256 randomAmount = bound(
+                _rerandomize(seed),
+                0,
+                remainingAmount
+            );
             joinSplitAmounts[i] = randomAmount;
             remainingAmount -= randomAmount;
         }